made all_habits and daily_habits conform to OOP
made new abstract class abstract_habit_list_fragment. each fragment that needs a list of habits from the DB can simply extend this and change the parseDataBaseUpdate function to serve their needs!
