+++ app/src/main/java/com/CMPUT301F21T19/habitappt/abstract_habit_list_fragment.java
@@ -1,0 +1,88 @@
+package com.CMPUT301F21T19.habitappt;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.EventListener;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.FirebaseFirestoreException;
+import com.google.firebase.firestore.QuerySnapshot;
+
+import java.util.ArrayList;
+
+public abstract class abstract_habit_list_fragment extends Fragment {
+
+    ListView habitListView;
+    ArrayAdapter<Habit> habitAdapter;
+    ArrayList<Habit> habitDataList;
+
+    View addHabitButton;
+    FirebaseFirestore db;
+
+    private View view;
+
+
+
+    //method for how a habit list fragment will parse a query.
+    public abstract void parseDataBaseUpdate(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable FirebaseFirestoreException error);
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.fragment_habits_list, container, false);
+
+        addHabitButton = view.findViewById(R.id.add_habit_button);
+        habitListView = view.findViewById(R.id.habit_list);
+
+        db = FirebaseFirestore.getInstance();
+
+        final CollectionReference collectionReference = db.collection("Default User");
+
+        habitDataList = new ArrayList<>();
+        habitAdapter = new HabitList(getContext(), habitDataList);
+        habitListView.setAdapter(habitAdapter);
+
+        habitListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                FragmentTransaction trans = getActivity().getSupportFragmentManager().beginTransaction();
+                trans.replace(R.id.main_container,new view_habit(habitDataList.get(position)));
+                trans.addToBackStack(null);
+                trans.commit();
+            }
+        });
+
+        addHabitButton.setOnClickListener( new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                new edit_habit().show(getActivity().getSupportFragmentManager(), "ADD");
+            }
+        });
+
+
+
+        collectionReference.addSnapshotListener(new EventListener<QuerySnapshot>() {
+            @Override
+            public void onEvent(@Nullable QuerySnapshot queryDocumentSnapshots, @Nullable FirebaseFirestoreException error) {
+                parseDataBaseUpdate(queryDocumentSnapshots,error);
+            }
+        });
+
+        return view;
+    }
+
+}
